Address distribute review feedback

+++ goldener/clusterize.py
@@ -330,7 +330,11 @@ class GoldClusterizer:
         self.batch_size = batch_size
         self.num_workers = num_workers
         self.allow_existing = allow_existing
-        self.distribute = distribute
+        if distribute:
+            raise NotImplementedError(
+                "Distributed processing is not implemented for GoldClusterizer."
+            )
+        self._distribute = distribute
         self.drop_table = drop_table
         self.max_batches = max_batches
         self.random_state = random_state
